attempt at licencing clarification
Orignal code was simply posted to Usenet.
ftp://ftp.freefriends.org/arnold/Awkstuff/xmlparser.awk

I had tried contacting Steve Collie around 2011, but the email bounced back.

I saw this code with a GPLv3 header tacked on, and the files here show that correspondence, (I wanted to use it in a BSD licenced project I was working on).

From here, I believe this code is simply in the public domain, but I am not a lawyer and cannot reccommend any approach to using this code in any licenced projects.

Signed-off-by: Isaac (.ike) Levy <ike@blackskyresearch.net>